requirements.txt
# Stuff easy to break with updates
torch==1.8.1
torchvision==0.9.1
pytorch-lightning==1.2.5
hydra-core==1.1.0.dev5
wandb==0.10.23
streamlit==0.79.0
datasets==1.6.2
transformers==4.5.1
# hydra-joblib-launcher

# Stable stuff usually backward compatible
dvc
python-dotenv
matplotlib
stqdm
